refactor(rest-api): modularize request handling into server and web-specific functions
This commit refactors the `rest-api` package by splitting request handling logic into separate modules for server and web environments. The `server_functions` and `web_functions` modules encapsulate environment-specific implementations, improving code organization and maintainability. Additionally, redundant code has been removed, and the `CorsLayer` configuration in `by-axum` has been updated for better flexibility.
